refactor(homepage): move navigator component off from haeder component
